Working system handling exceptions

+++ MobiBooking/Repository/UserRepository.cs
@@ -11,7 +11,7 @@ namespace MobiBooking.Models.Repository
 
         public UserRepository(BookingDbContext db)
         {
-            _db = db ?? throw new HttpResponseExceptionFilter(404, "Issue with connect to DataBase Context");
+            _db = db ?? throw new HttpResponseException(503, "Issue with connect to DataBase Context");
         }
 
         public IEnumerable<User> GetAll()
@@ -24,7 +24,7 @@ namespace MobiBooking.Models.Repository
             var user = _db.Users.FirstOrDefault(c => c.Id == id);
             if (user == null)
             {
-                throw new HttpResponseExceptionFilter(404, "Can't find user with this identifier!");
+                throw new HttpResponseException(404, "Can't find user with this identifier!");
             }
             return user;
         }
@@ -33,7 +33,7 @@ namespace MobiBooking.Models.Repository
         {
             if(user == null)
             {
-                throw new HttpResponseExceptionFilter(400, "Invalid sended data!");
+                throw new HttpResponseException(400, "Invalid sended data!");
             }
             _db.Users.Add(user);
             _db.SaveChanges();
@@ -57,7 +57,7 @@ namespace MobiBooking.Models.Repository
             }
             catch
             {
-                throw new HttpResponseExceptionFilter(400, "Invalid sended data!");
+                throw new HttpResponseException(400, "Invalid sended data!");
             }
 
             return user.Id;
@@ -65,15 +65,15 @@ namespace MobiBooking.Models.Repository
 
         public int Delete(int id)
         {
-            try
+            var user = _db.Users.FirstOrDefault(c => c.Id == id);
+            if(user == null)
             {
+                throw new HttpResponseException(404, "Can't find user with this identifier!");
+            }
+
                 _db.Users.Remove(new User() { Id = id });
                 _db.SaveChanges();
-            }
-            catch
-            {
-                throw new HttpResponseExceptionFilter(404, "Can't find user with this identifier!");
-            }
+            
 
             return id;
         }
